backend/requirements.txt
Flask==3.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.0
joblib==1.3.2
flask-cors==4.0.0
gunicorn==21.2.0
